Add emittery/legacy for Node.js 4 and 6 support
Fixes #9.
