+++ .travis.yml
@@ -1,3 +1,8 @@
 language: node_js
 node_js:
   - '8'
+  - '6'
+  - '4'
+install:
+  - npm install
+  - npm run build
