+++ readme.md
@@ -30,6 +30,9 @@ emitter.on('🦄', data => {
 emitter.emit('🦄', '🌈');
 ```
 
+### Node.js 4 and 6
+
+The above only works in Node.js 8 or newer. For older Node.js versions you can use `emittery/legacy`.
 
 ## API
